% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx.R
\name{read_xlsx_tool}
\alias{read_xlsx_tool}
\title{File, sheet, column Names Overview}
\usage{
read_xlsx_tool(file_paths)

read_xlsx_tool(file_paths)
}
\arguments{
\item{file_paths}{the path of the file which the data are to be read from.}
}
\value{
a tibble
}
\description{
Display xlsx file, worksheet,variable name in table to view data name
}
\details{
not done
}
\examples{
path <- system.file("extdata",package = "xlsxtools")
# full.names must be true, because openxlsx need read full path
file_paths <- list.files(path, full.names = T)
read_xlsx_tool(file_paths)

}
